자신의 글과 댓글만 수정,삭제하도록 변경

+++ src/main/java/com/mysite/review/controller/CommentController.java
@@ -2,6 +2,7 @@ package com.mysite.review.controller;
 
 import java.security.Principal;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.mysite.review.dto.CommentDTO;
 import com.mysite.review.dto.ReviewDTO;
@@ -51,28 +53,41 @@ public class CommentController {
 		
 	}
 	
+	@PreAuthorize("isAuthenticated()")
 	@GetMapping("/modify/{id}")
-	public String commentModify(CommentDTO commentDTO, @PathVariable("id") Long id) {
+	public String commentModify(CommentDTO commentDTO, @PathVariable("id") Long id, 
+			Principal principal) {
 		Comment comment = this.commentService.getComment(id);
+		if (!comment.getAuthor().getUsername().equals(principal.getName())) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+		}
 		commentDTO.setContent(comment.getContent());
 		return "comment_form";
 	}
 	
+	@PreAuthorize("isAuthenticated()")
 	@PostMapping("/modify/{id}")
 	public String commentModify(@Valid CommentDTO commentDTO, BindingResult bindingResult, 
-			@PathVariable("id") Long id) {
+			@PathVariable("id") Long id, Principal principal) {
 		if (bindingResult.hasErrors()) {
 			return "comment_form";
 		}
 		Comment comment = this.commentService.getComment(id);
+		if (!comment.getAuthor().getUsername().equals(principal.getName())) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+		}
 		this.commentService.modify(comment, commentDTO.getContent());
 		return String.format("redirect:/review/detail/%s#comment_%s", 
 				comment.getReview().getId(), comment.getId());
 	}
 	
+	@PreAuthorize("isAuthenticated()")
 	@GetMapping("/delete/{id}")
-	public String commentDelete(@PathVariable("id") Long id) {
+	public String commentDelete(@PathVariable("id") Long id, Principal principal) {
 		Comment comment = this.commentService.getComment(id);
+		if (!comment.getAuthor().getUsername().equals(principal.getName()) ) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+		}
 		this.commentService.delete(comment);
 		return String.format("redirect:/review/detail/%s",comment.getReview().getId());
 	}
